[✨ feat] 알림 아이콘 컴포넌트 분리 및 api 연동 #95
